pyproject.toml
[tool.poetry]
name = "find-uz"
version = "0.1.0"
description = ""
authors = ["Javakhir Shavkatov <zshavkatov51@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dockerpyze = "^2.2.0"
django = "^5.2"
djangorestframework = "^3.16.0"
uvicorn = "^0.34.1"
ngrok = "^1.4.0"
pillow = "^11.2.1"
whitenoise = "^6.9.0"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
windows-curses = "^2.4.1"
adrf = "^0.1.9"


[tool.poetry.group.dev.dependencies]
dockerpyze = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dpy]
name = "finduz"
entrypoint = ["python", "manage.py", "runserver"]
python = "3.12"
base-image = "python:3.12-slim"

